Updated files
Home now has a JS function that will determine whether to display if pneumothorax is detected or not. In future, this will be passed from back end but for now is just a constant for debugging.

Also a function added that will display the confidence level. Again in future this will be obtained from the back-end but is a local constant for now. Function is not fully working yet.

The new changes have created a number of bugs. These will be addressed in the near future.